wip on test portential filters api
